Extract shared getClosedItems logic into BaseGitHubProvider
Move parsing, auth, worker pool, and header construction into a shared
fetchClosedGitHubItems helper. Both providers now delegate to it,
ensuring consistent User-Agent headers and preventing drift.
